Split out resolvers
